Move all product feed dashboard components in a subfolder
