+++ _dev/apps/ui/src/components/product-feed-page/dashboard/feed-configuration/feed-configuration-card.spec.ts
