+++ _dev/apps/ui/src/components/product-feed-page/dashboard/submitted-products/submitted-products.vue
