+++ _dev/apps/ui/src/components/product-feed-page/dashboard/sync-history/sync-history.vue
