+++ _dev/apps/ui/src/components/product-feed-page/dashboard/verified-products/verified-products.ts
@@ -1,0 +1,5 @@
+export type VerificationStats = null|{
+  productsInCatalog: string;
+  verifiedProducts: string;
+  nonCompliantProducts: string;
+};
