--- _dev/apps/ui/src/components/product-feed-page/verified-products/verified-products.ts
@@ -1,19 +1,0 @@
-export type StatusCard = {
-  title: string;
-  description: string;
-  value: string | null;
-  reverseColors: boolean;
-  variant: string;
-  link?: {
-    to?: {
-      name: string;
-    };
-    href?: string;
-  };
-};
-
-export type VerificationStats = null|{
-  productsInCatalog: string;
-  verifiedProducts: string;
-  nonCompliantProducts: string;
-};
